#168 v2.02 업데이트 1. 검색 및 관광 페이지 최적화 2. 댓글: (1) 상태관리 단순 최적화 (2) 댓글 작성, 수정, 삭제 오류 수정 (3) 컴포넌트 props 최적화 3. 사용자 상태 관리 변경: Redux => useContext로 변경 4. 커스텀 훅 리팩토링: useCommentSubmit 5. 유틸 함수 리팩토링 6. image 임포트 수정

+++ src/features/comments/components/modal/DeleteModal.tsx
@@ -9,17 +9,17 @@ import { deleteComment } from "features/comments/api/firestoreUtils";
 import "./DeleteModal.css";
 
 interface DeleteProps {
+  role: string;
   comment_data: CommentType;
-  type: string;
 }
 
-const DeleteModal = ({ comment_data, type }: DeleteProps) => {
+const DeleteModal = ({ role, comment_data }: DeleteProps) => {
   const dispatch = useAppDispatch();
   const { createdAt, user_id, origin_id } = comment_data;
   const comment_id = createdAt + user_id;
 
   const clearModalInfo = () => {
-    dispatch(modalActions.clearModalInfo({ comment_id, type }));
+    dispatch(modalActions.toggleToastModal({ comment_id }));
   };
 
   useEffect(() => {
@@ -37,35 +37,37 @@ const DeleteModal = ({ comment_data, type }: DeleteProps) => {
   const deleteCommentHandler = async () => {
     if (!comment_id) return alert("댓글 정보가 없습니다.");
 
-    let api_state = "";
-    dispatch(modalActions.clearModalInfo({ type: "delete" }));
+    let api_state = "댓글을 삭제 중입니다.";
+    dispatch(modalActions.toggleToastModal({ api_state }));
 
     try {
       const isExisingOrigin = await deleteComment(origin_id, comment_id);
 
-      if (origin_id) {
-        /* 1. 내가 작성한 답글 삭제 (my) => myReply 상태만 업데이트 */
-        if (type === "my") {
-          dispatch(myReplyActions.deleteMyReply({ origin_id, comment_id }));
-        }
+      /* 1. 내가 방금 작성한 답글 삭제 (my) => myReply 상태만 업데이트 */
+      if (role === "my" && origin_id) {
+        dispatch(myReplyActions.deleteMyReply({ origin_id, comment_id }));
+      }
 
-        /* 답글 삭제 (reply) => replyComments, originComment 상태 업데이트 */
-        if (type === "reply") {
-          if (isExisingOrigin) {
-            dispatch(
-              originCommentActions.changeReplyCount({
-                type,
-                comment_id: origin_id,
-              })
-            );
-          }
-          dispatch(replyActions.deleteReply({ origin_id, comment_id }));
+      /* 답글 삭제 (reply) => replyComments, originComment 상태 업데이트 */
+      if (role === "reply" && origin_id) {
+        if (isExisingOrigin) {
+          dispatch(
+            originCommentActions.changeReplyCount({
+              type: "delete",
+              comment_id: origin_id,
+            }),
+          );
         }
-      } else {
+        dispatch(replyActions.deleteReply({ origin_id, comment_id }));
+      }
+
+      /* 오리지널(최상위) 댓글 삭제 */
+      if (role === "origin") {
         dispatch(replyActions.deleteReply({ origin_id: comment_id }));
         dispatch(myReplyActions.deleteMyReply({ origin_id: comment_id }));
         dispatch(originCommentActions.deleteComment({ comment_id }));
       }
+
       api_state = "댓글을 삭제했습니다.";
     } catch (error: any) {
       api_state = "오류가 발생했습니다.";
